x/cron/types/messages_schedule_test.go
package types

import (
	"github.com/neutron-org/neutron/testutil/cron/sample"
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateSchedule_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateSchedule
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateSchedule{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateSchedule{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateSchedule_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateSchedule
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateSchedule{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateSchedule{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteSchedule_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteSchedule
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteSchedule{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteSchedule{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
